requirements.txt
certifi==2019.6.16
chardet==3.0.4
Django==2.2
idna==2.8
mysqlclient==1.4.2.post1
Naked==0.1.31
pycryptodome==3.8.2
python-crontab==2.3.6
python-dateutil==2.8.0
pytz==2019.1
PyYAML==5.1.1
requests==2.21.0
shellescape==3.4.1
six==1.12.0
sqlparse==0.3.0
urllib3==1.24.3
